added formatter for main page which format data to component format

+++ Weather App/Components/Weather Detailized Component/Atmospheric Events Card View/AtmosphericEventsCardViewData.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct AtmosphericEventsCardViewData {
-    let rainValue: Double
-    let humidityValue: Double
+    let rainValue: Int
+    let humidityValue: Int
     let windValue: Double
 }
